work_with_database/requirements.txt
django~=4.2.7
psycopg2-binary

django-autoslug~=1.9.9